ucla cs143
Basically there are 2 projects.
Project 1 is to Built a movie search engine supported by a relational
database system.The system managed all of its data at the back-end in a
MySQL database and provided a Web interface to the users at the
front-end.
Project 2 is to Created the multi-level dense index according to the
key value when loading or adding data to database to reduced disk I/O
for query.